bot_requirements.txt
# Dependencies for discord_bot.py (run separately from GitHub Actions)
discord.py>=2.0
google-generativeai
requests
